Update Wing_v2.py
I added more comments, corrected the aoa in 202,203 to radians,since the np.cos and np.sin takes the input as radians. Made changes in input of biot_savart function and I think it is correct now. But for that the range of j in for loop is changed, so half of the  values would be 0 as of now.